requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.6.4
chromadb==0.5.5
langchain==0.2.11
langchain-core==0.2.23
langchain-community==0.2.10
google-generativeai==0.7.2
firecrawl==1.4.0
pincone
